+++ packages/interaction/CHANGELOG.md
@@ -8,3 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - upgrade `@bigtest/convergence` to `0.4.0`
+
+### Fixed
+
+- `isPresent` property returns false when the root does not exist